Demonstrate how to implement hierarchical logging
