+++ substratevm/src/com.oracle.objectfile/src/com/oracle/objectfile/elf/dwarf/DwarfSections.java
@@ -26,6 +26,15 @@
 
 package com.oracle.objectfile.elf.dwarf;
 
+import java.nio.ByteOrder;
+import java.nio.file.Path;
+import java.util.HashMap;
+import java.util.LinkedList;
+import java.util.List;
+import java.util.Map;
+
+import org.graalvm.compiler.debug.DebugContext;
+
 import com.oracle.objectfile.debugentry.ClassEntry;
 import com.oracle.objectfile.debugentry.DirEntry;
 import com.oracle.objectfile.debugentry.FileEntry;
@@ -34,13 +43,6 @@ import com.oracle.objectfile.debugentry.StringTable;
 import com.oracle.objectfile.debuginfo.DebugInfoProvider;
 import com.oracle.objectfile.debuginfo.DebugInfoProvider.DebugFrameSizeChange;
 import com.oracle.objectfile.elf.ELFMachine;
-
-import java.nio.ByteOrder;
-import java.nio.file.Path;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Map;
 
 /**
  * A class that models the debug info in an organization that facilitates generation of the required
@@ -90,24 +92,19 @@ public class DwarfSections {
     public static final int DW_AT_hi_pc = 0x12;
     public static final int DW_AT_language = 0x13;
     public static final int DW_AT_external = 0x3f;
-    @SuppressWarnings("unused")
-    public static final int DW_AT_return_addr = 0x2a;
-    @SuppressWarnings("unused")
-    public static final int DW_AT_frame_base = 0x40;
+    @SuppressWarnings("unused") public static final int DW_AT_return_addr = 0x2a;
+    @SuppressWarnings("unused") public static final int DW_AT_frame_base = 0x40;
     /*
      * define all the Dwarf attribute forms we need for our DIEs
      */
     public static final int DW_FORM_null = 0x0;
-    @SuppressWarnings("unused")
-    private static final int DW_FORM_string = 0x8;
+    @SuppressWarnings("unused") private static final int DW_FORM_string = 0x8;
     public static final int DW_FORM_strp = 0xe;
     public static final int DW_FORM_addr = 0x1;
     public static final int DW_FORM_data1 = 0x0b;
     public static final int DW_FORM_data4 = 0x6;
-    @SuppressWarnings("unused")
-    public static final int DW_FORM_data8 = 0x7;
-    @SuppressWarnings("unused")
-    public static final int DW_FORM_block1 = 0x0a;
+    @SuppressWarnings("unused") public static final int DW_FORM_data8 = 0x7;
+    @SuppressWarnings("unused") public static final int DW_FORM_block1 = 0x0a;
     public static final int DW_FORM_flag = 0xc;
 
     /*
@@ -121,8 +118,7 @@ public class DwarfSections {
     /*
      * DW_FORM_flag attribute values
      */
-    @SuppressWarnings("unused")
-    public static final byte DW_FLAG_false = 0;
+    @SuppressWarnings("unused") public static final byte DW_FLAG_false = 0;
     public static final byte DW_FLAG_true = 1;
     /*
      * value for DW_AT_language attribute with form DATA1
@@ -134,21 +130,16 @@ public class DwarfSections {
      *
      * not needed until we make functions members
      */
-    @SuppressWarnings("unused")
-    public static final byte DW_ACCESS_public = 1;
-    @SuppressWarnings("unused")
-    public static final byte DW_ACCESS_protected = 2;
-    @SuppressWarnings("unused")
-    public static final byte DW_ACCESS_private = 3;
+    @SuppressWarnings("unused") public static final byte DW_ACCESS_public = 1;
+    @SuppressWarnings("unused") public static final byte DW_ACCESS_protected = 2;
+    @SuppressWarnings("unused") public static final byte DW_ACCESS_private = 3;
 
     /*
      * others not yet needed
      */
-    @SuppressWarnings("unused")
-    public static final int DW_AT_type = 0; // only present for non-void
+    @SuppressWarnings("unused") public static final int DW_AT_type = 0; // only present for non-void
                                                                         // functions
-    @SuppressWarnings("unused")
-    public static final int DW_AT_accessibility = 0;
+    @SuppressWarnings("unused") public static final int DW_AT_accessibility = 0;
 
     /*
      * CIE and FDE entries
@@ -156,36 +147,28 @@ public class DwarfSections {
 
     /* full byte/word values */
     public static final int DW_CFA_CIE_id = -1;
-    @SuppressWarnings("unused")
-    public static final int DW_CFA_FDE_id = 0;
+    @SuppressWarnings("unused") public static final int DW_CFA_FDE_id = 0;
 
     public static final byte DW_CFA_CIE_version = 1;
 
     /* values encoded in high 2 bits */
     public static final byte DW_CFA_advance_loc = 0x1;
     public static final byte DW_CFA_offset = 0x2;
-    @SuppressWarnings("unused")
-    public static final byte DW_CFA_restore = 0x3;
+    @SuppressWarnings("unused") public static final byte DW_CFA_restore = 0x3;
 
     /* values encoded in low 6 bits */
     public static final byte DW_CFA_nop = 0x0;
-    @SuppressWarnings("unused")
-    public static final byte DW_CFA_set_loc1 = 0x1;
+    @SuppressWarnings("unused") public static final byte DW_CFA_set_loc1 = 0x1;
     public static final byte DW_CFA_advance_loc1 = 0x2;
     public static final byte DW_CFA_advance_loc2 = 0x3;
     public static final byte DW_CFA_advance_loc4 = 0x4;
-    @SuppressWarnings("unused")
-    public static final byte DW_CFA_offset_extended = 0x5;
-    @SuppressWarnings("unused")
-    public static final byte DW_CFA_restore_extended = 0x6;
-    @SuppressWarnings("unused")
-    public static final byte DW_CFA_undefined = 0x7;
-    @SuppressWarnings("unused")
-    public static final byte DW_CFA_same_value = 0x8;
+    @SuppressWarnings("unused") public static final byte DW_CFA_offset_extended = 0x5;
+    @SuppressWarnings("unused") public static final byte DW_CFA_restore_extended = 0x6;
+    @SuppressWarnings("unused") public static final byte DW_CFA_undefined = 0x7;
+    @SuppressWarnings("unused") public static final byte DW_CFA_same_value = 0x8;
     public static final byte DW_CFA_register = 0x9;
     public static final byte DW_CFA_def_cfa = 0xc;
-    @SuppressWarnings("unused")
-    public static final byte DW_CFA_def_cfa_register = 0xd;
+    @SuppressWarnings("unused") public static final byte DW_CFA_def_cfa_register = 0xd;
     public static final byte DW_CFA_def_cfa_offset = 0xe;
 
     private ByteOrder byteOrder;
@@ -326,6 +309,7 @@ public class DwarfSections {
      * 
      * @param debugInfoProvider provider instance passed by ObjectFile client
      */
+    @SuppressWarnings("try")
     public void installDebugInfo(DebugInfoProvider debugInfoProvider) {
         /*
          * DebugTypeInfoProvider typeInfoProvider = debugInfoProvider.typeInfoProvider(); for
@@ -337,7 +321,7 @@ public class DwarfSections {
          */
         uniqueDebugString("");
 
-        debugInfoProvider.codeInfoProvider().forEach(debugCodeInfo -> {
+        debugInfoProvider.codeInfoProvider().forEach(debugCodeInfo -> debugCodeInfo.debugContext((debugContext) -> {
             /*
              * primary file name and full method name need to be written to the debug_str section
              */
@@ -351,7 +335,9 @@ public class DwarfSections {
             int lo = debugCodeInfo.addressLo();
             int hi = debugCodeInfo.addressHi();
             int primaryLine = debugCodeInfo.line();
+
             Range primaryRange = new Range(fileName, filePath, className, methodName, paramNames, returnTypeName, stringTable, lo, hi, primaryLine);
+            debugContext.log("PrimaryRange %s %s.%s", fileName, className, methodName);
             /*
              * System.out.format("arange: [0x%08x,0x%08x) %s %s::%s(%s) %s\n", lo, hi,
              * returnTypeName, className, methodName, paramNames, fileName); create an infoSection
@@ -373,8 +359,11 @@ public class DwarfSections {
                  */
                 Range subRange = new Range(fileNameAtLine, filePathAtLine, classNameAtLine, methodNameAtLine, "", "", stringTable, loAtLine, hiAtLine, line, primaryRange);
                 addSubRange(primaryRange, subRange);
+                try (DebugContext.Scope s = debugContext.scope("Subranges")) {
+                    debugContext.log("SubRange %s %s %s.%s", fileNameAtLine, filePathAtLine, classNameAtLine, methodNameAtLine);
+                }
             });
-        });
+        }));
         /*
          * DebugDataInfoProvider dataInfoProvider = debugInfoProvider.dataInfoProvider(); for
          * (DebugDataInfo debugDataInfo : dataInfoProvider) { install details of heap elements
